requirements.txt
fastapi==0.96.0
uvicorn==0.22.0
jinja2==3.1.2
httpx==0.24.1
python-dotenv==1.0.0
pytest==7.3.1  # For testing (optional)
